Remove sickle mod by moving to builtin/assets
Note: this required the removal of the "assets" feature as now the fonts plugin is necessary
