+++ examples/game_menu/src/main.rs
@@ -6,7 +6,7 @@ use bevy::window::{PresentMode, PrimaryWindow, WindowTheme};
 use bevy_cobweb::prelude::*;
 use bevy_cobweb_ui::builtin::widgets::radio_button::{RadioButtonBuilder, RadioButtonManager};
 use bevy_cobweb_ui::prelude::*;
-use bevy_cobweb_ui::sickle::prelude::*;
+use bevy_cobweb_ui::sickle_ext::prelude::*;
 
 //-------------------------------------------------------------------------------------------------------------------
