+++ examples/help_text/src/main.rs
@@ -8,7 +8,7 @@
 use bevy::prelude::*;
 use bevy::window::WindowTheme;
 use bevy_cobweb_ui::prelude::*;
-use bevy_cobweb_ui::sickle::prelude::*;
+use bevy_cobweb_ui::sickle_ext::prelude::*;
 
 //-------------------------------------------------------------------------------------------------------------------
