+++ examples/localization/src/main.rs
@@ -5,7 +5,7 @@ use bevy::window::WindowTheme;
 use bevy_cobweb::prelude::*;
 use bevy_cobweb_ui::builtin::widgets::radio_button::{RadioButtonBuilder, RadioButtonManager};
 use bevy_cobweb_ui::prelude::*;
-use bevy_cobweb_ui::sickle::ui_builder::*;
+use bevy_cobweb_ui::sickle_ext::ui_builder::*;
 
 //-------------------------------------------------------------------------------------------------------------------
