+++ src/builtin/assets/fonts/MaterialIcons-Regular.codepoints
