+++ src/builtin/assets/fonts/MaterialIcons-Regular.ttf
