+++ src/builtin/widgets/radio_button/mod.rs
@@ -3,7 +3,7 @@ use bevy_cobweb::prelude::*;
 
 //use crate::load_embedded_scene_file;
 use crate::prelude::*;
-use crate::sickle::prelude::*;
+use crate::sickle_ext::prelude::*;
 
 //-------------------------------------------------------------------------------------------------------------------
