+++ src/sickle_ext/control.rs
@@ -5,8 +5,8 @@ use serde::{Deserialize, Serialize};
 use smol_str::SmolStr;
 
 use crate::prelude::*;
-use crate::sickle::prelude::DynamicStyle;
-use crate::sickle::theme::dynamic_style::DynamicStyleStopwatch;
+use crate::sickle_ext::prelude::DynamicStyle;
+use crate::sickle_ext::theme::dynamic_style::DynamicStyleStopwatch;
 
 //-------------------------------------------------------------------------------------------------------------------
 
@@ -138,7 +138,7 @@ impl Instruction for ControlRoot
 ///
 /// Use this if you want values on the entity to respond to interactions on other parts of the widget, or if
 /// you want different values to be applied depending on the widget's
-/// [`PseudoStates`](crate::sickle::prelude::PseudoState).
+/// [`PseudoStates`](crate::sickle_ext::prelude::PseudoState).
 ///
 /// Values in a multi-entity widget can be controlled with the [`Themed`], [`Responsive`], and [`Animated`]
 /// loadables.
