+++ src/sickle_ext/interaction_ext.rs
@@ -3,8 +3,8 @@ use bevy_cobweb::prelude::*;
 use serde::{Deserialize, Serialize};
 
 use crate::prelude::*;
-use crate::sickle::prelude::*;
-use crate::sickle::theme::pseudo_state::{PseudoState, PseudoStates};
+use crate::sickle_ext::prelude::*;
+use crate::sickle_ext::theme::pseudo_state::{PseudoState, PseudoStates};
 
 //-------------------------------------------------------------------------------------------------------------------
