+++ src/sickle_ext/mod.rs
@@ -7,7 +7,6 @@ mod control;
 mod control_attributes;
 mod control_loadable_registration;
 mod control_map;
-mod default_assets;
 mod interaction_ext;
 mod plugin;
 mod pseudo_states_ext;
@@ -17,7 +16,15 @@ pub use control::*;
 pub use control_attributes::*;
 pub use control_loadable_registration::*;
 pub(crate) use control_map::*;
-pub(crate) use default_assets::*;
 pub use interaction_ext::*;
 pub(crate) use plugin::*;
 pub use pseudo_states_ext::*;
+pub use sickle_macros::*;
+pub use sickle_math::*;
+pub use sickle_ui_scaffold::*;
+
+pub mod prelude {
+    pub use sickle_macros::*;
+    pub use sickle_math::*;
+    pub use sickle_ui_scaffold::prelude::*;
+}
