+++ src/sickle_ext/plugin.rs
@@ -1,9 +1,12 @@
 use bevy::prelude::*;
 
 use crate::prelude::*;
-use crate::sickle::ease::Ease;
-use crate::sickle::theme::pseudo_state::PseudoState;
-use crate::sickle::theme::style_animation::{AnimationConfig, AnimationLoop, AnimationSettings};
+use crate::sickle_ext::ease::Ease;
+use crate::sickle_ext::theme::pseudo_state::PseudoState;
+use crate::sickle_ext::theme::style_animation::{AnimationConfig, AnimationLoop, AnimationSettings};
+use bevy::prelude::*;
+use flux_interaction::FluxInteractionPlugin;
+use theme::ThemePlugin;
 
 //-------------------------------------------------------------------------------------------------------------------
 
@@ -13,12 +16,15 @@ impl Plugin for SickleExtPlugin
 {
     fn build(&self, app: &mut App)
     {
+        app.add_plugins((
+            FluxInteractionPlugin,
+            ThemePlugin,
+        ));
         app.register_type::<Ease>()
             .register_type::<PseudoState>()
             .register_type::<AnimationSettings>()
             .register_type::<AnimationConfig>()
             .register_type::<AnimationLoop>()
-            .add_plugins(SickleUiDefaultAssetsPlugin)
             .add_plugins(ControlPlugin)
             .add_plugins(ControlMapPlugin)
             .add_plugins(UiInteractionExtPlugin)
