+++ src/sickle_ext/pseudo_states_ext.rs
@@ -2,9 +2,9 @@ use bevy::prelude::*;
 use bevy_cobweb::prelude::*;
 
 use crate::prelude::*;
-use crate::sickle::prelude::UiBuilder;
-use crate::sickle::theme::pseudo_state::PseudoState;
-use crate::sickle::ui_commands::ManagePseudoStateExt;
+use crate::sickle_ext::prelude::UiBuilder;
+use crate::sickle_ext::theme::pseudo_state::PseudoState;
+use crate::sickle_ext::ui_commands::ManagePseudoStateExt;
 
 //-------------------------------------------------------------------------------------------------------------------
