+++ src/ui_bevy/ui_ext/component_wrappers.rs
@@ -4,7 +4,7 @@ use bevy_cobweb::prelude::*;
 use serde::{Deserialize, Serialize};
 
 use crate::prelude::*;
-use crate::sickle::lerp::Lerp;
+use crate::sickle_ext::lerp::Lerp;
 
 //-------------------------------------------------------------------------------------------------------------------
