+++ src/ui_bevy/ui_ext/style_wrappers.rs
@@ -3,7 +3,7 @@ use bevy_cobweb::prelude::*;
 use serde::{Deserialize, Serialize};
 
 use crate::prelude::*;
-use crate::sickle::lerp::Lerp;
+use crate::sickle_ext::lerp::Lerp;
 
 //-------------------------------------------------------------------------------------------------------------------
